4. Разделила код на модули
