+++ js/data.js
@@ -1,0 +1,13 @@
+const NAMES = ['Рэйчел Грин', 'Моника Геллер', 'Фиби Буффе', 'Джоуи Триббиани', 'Чендлер Бинг', 'Елена Гилберт', 'Кэтрин Пирс', 'Стефан Сальваторе', 'Дэймон Сальваторе', 'Джереми Гилберт', 'Дженна Соммерс', 'Бонни Беннет', 'Кэролайн Форбс', 'Мэтт Донован', 'Викки Донован', 'Тайлер Локвуд', 'Аларик Зальцман', 'Клаус Майклсон', 'Энзо Сент-Джон', 'Элайджа Майклсон', 'Коул Майклсон', 'Ребекка Майклсон', 'Финн Майклсон', 'Лив Паркер', 'Люк Паркер'];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'];
+
+const SIMILAR_PHOTO_DESCRIPTION_COUNT = 25;
+
+export {NAMES, MESSAGES, SIMILAR_PHOTO_DESCRIPTION_COUNT};
